pyproject.toml
[project]
name = "ipymlwidgets"
version = "0.1.0"
description = ""
authors = [
    { name = "Benedict Wilkins", email = "benedict.wilkins@sony.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pillow",
    "ipywidgets",
    "ipycanvas",
    "ipykernel",
]

[project.optional-dependencies]
dev = [ 
    "ruff",       # must use this!
    "pre-commit", # must use this!
    "pytest",     # must use this!
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"datautils.fonts" = ["*.ttf"]

# ====== pytest config ====== # 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -v"
testpaths = ["tests"]

# ====== ruff config ====== # 

[tool.ruff]
# Exclude common directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
# we care about docstrings, flakes (undefined vars, etc.), and some lint rules
select = ["D", "F", "E4", "E7", "E9"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google" 

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D100"]        # Ignore missing docstrings in __init__.py
"tests/**/*.py" = ["D"]         # Ignore all pydocstyle rules in tests
"**/*.ipynb" = ["D", "E", "F"]  # Ignore all rules in prototyping notebooks... these are a mess anyway!
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
docstring-code-format = false
docstring-code-line-length = "dynamic"

# ====== uv config ====== # 

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.7",
    "pre-commit",
    "pytest",
]
#no-build-isolation-package = ["torch_scatter"]

# Recommended to use `[uv](https://docs.astral.sh/uv/)` to install this package.
# `uv` will automatically install pytorch with CUDA. 
[tool.uv.sources]
torch = [
  { index = "pytorch-cu118" }
]
torchvision = [
  { index = "pytorch-cu118" }
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true